Fixed the async example
- Added tokio as a dev-dependency
- Use "current_thread" tokio runtime

The current_thread tokio runtime ensures that the publisher has finished
publishing before the subscribers start to receive. The example may fail
in the default thread_pool runtime as there is a race condition between
the publisher and subscriber, where the subscriber may get all the
messages instead of just the last 10.