Set missed items size and upgrade futures
Add a `missed_items_size: usize` argument to the `bounded` method so
that the user can designate how much items should be ignored when the
writer overflows a slow reading client.

This is handy in high throughput scenarios where we want to ignore all
previous messages and send the last one in the buffer to the slow
readers.