Refactor wakers to use piper::Event to notify pending streams an item is ready
Stream Trait is similar to the Future Trait in that it requires internal logic
to let the executor know when the stream's state changes from Poll::Pending to
Poll::Ready.

In the original implementation mpsc channel and AtomicWakers were used. The
AtomicWaker was used to contain the future's context and wake the task, while
the channel was used to send AtomicWakers from Subscribers to the Publisher.
After publishing a new item the Publisher collected all the AtomicWakers and
called the wake method.

This refactor replaces that logic with piper::Event which is awaited in the
Stream::poll_next method until the Publisher signals that event by calling the
Event::notify method.

An additional close method was added to channel::Sender which explicitly let's
all Receivers and Subscribers know that the Sender or Publisher has stopped
publishing items.